Move inria pedestrian code into a separate folder
