+++ dbcollection/datasets/inria/pedestrian/README.rst
@@ -21,6 +21,6 @@ Data structure
 Disclaimer
 ----------
 
-All rights reserved to the original creators of ``CIFAR-100``.
+All rights reserved to the original creators of ``INRIA Pedestrian``.
 
 For information about this dataset and its terms of use, please see the `original website <http://pascal.inrialpes.fr/data/human//>`_.
